sentinel-shield/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.3.0
PyJWT==2.12.1
requests==2.33.1
SQLAlchemy==2.0.49
typing_extensions==4.15.0
ua-parser==1.0.2
ua-parser-builtins==202603
urllib3==2.6.3
user-agents==2.2.0
Werkzeug==3.1.8
